# src/main/java/ru/liga/application/domain/dto/TaskDto.java
package ru.liga.application.domain.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@Builder
@Data
public class TaskDto {
    private long employeeId;
    private String uuid;
    private boolean completed;
    private String description;
}